requirements.txt
asgiref==3.6.0
Django==4.1.5
gunicorn==20.1.0
psycopg2==2.9.5
python-decouple==3.7
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0